# docs/cmdb-api-spec.yaml
openapi: 3.0.1
info:
  title: Chelas Movies Database
  description: Movies API using IMDB API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - description: Localhost server for testing API
    url: http://localhost:8080/
...
paths:
  /api/users:
    post:
      tags:
      - Users
      summary: adds a user
      description: Adds a user to the system
      operationId: createUser
      requestBody:
        description: User to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
        required: false
      responses:
        201:
          description: user created
          content:
            application/json:
              schema:
                required:
                - token
                type: object
                properties:
                  token:
                    type: string
                    format: uuid
        400:
          description: Invalid input, object invalid
          content: {}
      x-codegen-request-body-name: user
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    NewUser:
      required:
      - name
      type: object
      properties:
        name:
          type: string
          example: joao
...